refactor : shared/common 폴더 생성
