+++ src/App.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { BrowserRouter } from 'react-router-dom';
 import { ThemeProvider } from '@emotion/react';
-import { Router } from '@/common/components/Router';
+import { Router } from '@/shared/common/components/Router';
 import { CategoryProvider } from '@/features/Situation/CategoryContext';
-import { Layout } from '@/common/components/Layout';
+import { Layout } from '@/shared/common/components/Layout';
 import { GlobalStyle, theme } from '@/styles';
-import { Toast } from '@/common/components/Toast';
-import { BottomSheet } from '@/common/components/BottomSheet';
+import { Toast } from '@/shared/common/components/Toast';
+import { BottomSheet } from '@/shared/common/components/BottomSheet';
 
 const queryClient = new QueryClient();
