+++ src/features/MusicSearch/components/MusicSearchInput.tsx
@@ -1,8 +1,8 @@
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 import { SetStateAction, KeyboardEvent, Dispatch } from 'react';
-import { Input } from '@/common/components/Input';
-import { SvgIcon } from '@/common/components/SvgIcon';
+import { Input } from '@/shared/common/components/Input';
+import { SvgIcon } from '@/shared/common/components/SvgIcon';
 import { theme } from '@/styles';
 
 interface Props {
