+++ src/shared/common/components/Footer/Footer.tsx
