+++ src/shared/common/components/Footer/index.ts
