+++ src/shared/common/components/Layout/index.ts
