+++ src/shared/common/components/Modal/index.ts
