+++ src/shared/common/components/SvgIcon/SvgIcon.tsx
