+++ src/shared/common/components/SvgIcon/assets.ts
