+++ src/shared/common/components/SvgIcon/index.ts
