+++ src/shared/common/hooks/index.ts
