+++ src/shared/common/hooks/useClickOutside.ts
