+++ src/shared/common/hooks/useOpen.ts
